refactor: Port 변경
TrimPort에서 GetTrimsPort로 변겨
